about half way there for pointer return args:
if an argument to a method is a pointer, we assume it's intended to be
available on the return, so the return type of any method which has one
or more pointer arguments will be an array. the first value in the returned
array is the return value of the method. the subsequent values in the array
will be the n-1th arguments.

right now, we don't set the values of pointer arguments going in.
additionally, these values are optional and don't need to be provided.

eventually, we will set the ingoing values.

currently, only char, int and id arguments are returned in this manner.
additional types will be added once this proves to be a workable technique.

for an example, see:

[NSWorkspace getFileSystemInfoForPath:isRemovable:isWritable:isUnmountable:description:type:]

also, readded _sprintf() to objects so we can see what they look like.

--HG--
extra : convert_revision : svn%3A2b5a5965-8a46-0410-aa19-a76fe43533fc/trunk%403021